Fix gcc compil errors/warnings in tests

+++ modules/utils/test/platform.cpp
@@ -31,7 +31,7 @@ TEST(filesystem_ops,regression) {
     lv::filterFilePaths(vsFiles2,(std::vector<std::string>{}),(std::vector<std::string>{".txt"}));
     EXPECT_EQ(vsFiles2,(std::vector<std::string>{sDirPath+"test1.txt"}));
     EXPECT_EQ(lv::getSubDirsFromDir(sDirPath),(std::vector<std::string>{sDirPath+"subdir1",sDirPath+"subdir2"}));
-    EXPECT_GT(lv::getCurrentPhysMemBytesUsed(),0);
+    EXPECT_GT(lv::getCurrentPhysMemBytesUsed(),size_t(0));
 }
 
 namespace {
@@ -42,11 +42,11 @@ namespace {
 TYPED_TEST_CASE(AlignedMemAllocator_fixture,AlignedMemAllocator_types);
 TYPED_TEST(AlignedMemAllocator_fixture,regression) {
     std::vector<TypeParam,lv::AlignedMemAllocator<TypeParam,16>> vVec16a(100);
-    EXPECT_EQ(((uintptr_t)vVec16a.data()%16),0);
+    EXPECT_EQ(((uintptr_t)vVec16a.data()%16),size_t(0));
     ASSERT_EQ(vVec16a[0],TypeParam(0));
     ASSERT_TRUE(std::equal(vVec16a.begin()+1,vVec16a.end(),vVec16a.begin()));
     std::vector<TypeParam,lv::AlignedMemAllocator<TypeParam,32>> vVec32a(100);
-    EXPECT_EQ(((uintptr_t)vVec32a.data()%32),0);
+    EXPECT_EQ(((uintptr_t)vVec32a.data()%32),size_t(0));
     ASSERT_EQ(vVec32a[0],TypeParam(0));
     ASSERT_TRUE(std::equal(vVec32a.begin()+1,vVec32a.end(),vVec32a.begin()));
 }
